type-c-service/wrapper: Rename and flatten module
Flatten `wrapper/proxy` into `controller` as the previous `wrapper`
module no longer contains any code of its own and `controller` is a
better name than `wrapper`.
