+++ examples/rt685s-evk/src/bin/type_c.rs
@@ -27,13 +27,13 @@ use type_c_interface::port::{ControllerId, Device};
 use type_c_interface::service::event::PortEvent as ServicePortEvent;
 use type_c_service::bridge::Bridge;
 use type_c_service::bridge::event_receiver::EventReceiver as BridgeEventReceiver;
-use type_c_service::driver::tps6699x::{self as tps6699x_drv};
-use type_c_service::service::{EventReceiver as ServiceEventReceiver, Service};
-use type_c_service::wrapper::proxy::PowerProxyDevice;
-use type_c_service::wrapper::proxy::event_receiver::{
+use type_c_service::controller::Port;
+use type_c_service::controller::event_receiver::{
     EventReceiver as PortEventReceiver, InterruptReceiver as _, PortEventSplitter,
 };
-use type_c_service::wrapper::proxy::state::SharedState;
+use type_c_service::controller::state::SharedState;
+use type_c_service::driver::tps6699x::{self as tps6699x_drv};
+use type_c_service::service::{EventReceiver as ServiceEventReceiver, Service};
 
 extern crate rt685s_evk_example;
 
@@ -50,7 +50,7 @@ bind_interrupts!(struct Irqs {
 type SharedStateType = Mutex<GlobalRawMutex, SharedState>;
 type PortType = Mutex<
     GlobalRawMutex,
-    PowerProxyDevice<
+    Port<
         'static,
         Tps6699xMutex<'static>,
         SharedStateType,
@@ -236,7 +236,7 @@ async fn main(spawner: Spawner) {
     let port1_shared_state = PORT1_SHARED_STATE.init(Mutex::new(SharedState::new()));
 
     static PORT0: StaticCell<PortType> = StaticCell::new();
-    let port0 = PORT0.init(Mutex::new(PowerProxyDevice::new(
+    let port0 = PORT0.init(Mutex::new(Port::new(
         "PD0",
         Default::default(),
         LocalPortId(0),
@@ -248,7 +248,7 @@ async fn main(spawner: Spawner) {
     )));
 
     static PORT1: StaticCell<PortType> = StaticCell::new();
-    let port1 = PORT1.init(Mutex::new(PowerProxyDevice::new(
+    let port1 = PORT1.init(Mutex::new(Port::new(
         "PD1",
         Default::default(),
         LocalPortId(1),
