+++ examples/std/src/bin/type_c/ucsi.rs
@@ -28,13 +28,12 @@ use type_c_interface::service::context::Context;
 use type_c_interface::service::event::{PortEvent as ServicePortEvent, PortEventData as ServicePortEventData};
 use type_c_service::bridge::Bridge;
 use type_c_service::bridge::event_receiver::EventReceiver as BridgeEventReceiver;
+use type_c_service::controller::event::Event as PortEvent;
+use type_c_service::controller::event_receiver::InterruptReceiver as _;
+use type_c_service::controller::event_receiver::{EventReceiver as PortEventReceiver, PortEventSplitter};
+use type_c_service::controller::state::SharedState;
 use type_c_service::service::config::Config;
 use type_c_service::service::{EventReceiver as ServiceEventReceiver, Service};
-use type_c_service::wrapper::proxy::PowerProxyDevice;
-use type_c_service::wrapper::proxy::event::Event as PortEvent;
-use type_c_service::wrapper::proxy::event_receiver::InterruptReceiver as _;
-use type_c_service::wrapper::proxy::event_receiver::{EventReceiver as PortEventReceiver, PortEventSplitter};
-use type_c_service::wrapper::proxy::state::SharedState;
 
 const CHANNEL_CAPACITY: usize = 4;
 const NUM_PD_CONTROLLERS: usize = 2;
@@ -292,7 +291,7 @@ async fn task(spawner: Spawner) {
     let port_interrupt_sender_0 = port_interrupt_channel_0.dyn_sender();
 
     static PORT0: StaticCell<PortType> = StaticCell::new();
-    let port0 = PORT0.init(Mutex::new(PowerProxyDevice::new(
+    let port0 = PORT0.init(Mutex::new(Port::new(
         "PD0",
         Default::default(),
         LocalPortId(0),
@@ -337,7 +336,7 @@ async fn task(spawner: Spawner) {
     let port_interrupt_sender_1 = port_interrupt_channel_1.dyn_sender();
 
     static PORT1: StaticCell<PortType> = StaticCell::new();
-    let port1 = PORT1.init(Mutex::new(PowerProxyDevice::new(
+    let port1 = PORT1.init(Mutex::new(Port::new(
         "PD1",
         Default::default(),
         LocalPortId(0),
