--- type-c-service/src/wrapper/mod.rs
@@ -1,1 +1,0 @@
-pub mod proxy;
